Architecture redesign
Switch from a design focused around streaming, with a classes that interpret
events as they are requested, to an in-memory format-independent representation
of a song, with one class to handle reading and writing a file format.

Rdos .raw format has been rewritten for the new architecture, as have the
matching tests and example programs, but the other format handlers have not
yet been converted.