dro2txt: Support OPL3/second OPL2 output
A lowercase "b" is now displayed next to the channel (or percussive instrument)
if it is for the second OPL2 chip, or the upper set of OPL3 registers.  OPL3
specific registers (like the 4-OP ones) are now also identified in the output.

The global state is also now printed on every change rather than only when
notes are audible, to avoid missing changes made while no notes are playing.