Add oplNormalisePerc() function to handle formats with crossed operators
Some file formats load the operator for the percussive instrument from the
modulator field in the music file, even when those parameters get placed into
the carrier operator instead.  This function swaps the operators as required
so that on return, modulator-only instrument parameters are always loaded from
the modulator fields, and carrier-only instrument parameters are always loaded
from the carrier fields.