Allow tests to skip isInstance checks against given formats
This allows certain formats to flag that they are always misdetected as another
format, so that check is bypassed as there is nothing we can do about it.  This
can happen when two otherwise identical file formats have a different tempo, as
in the case for mus-dmx in Raptor vs the other games.