decode-opl: Handle abrupt sustain/release changes as volume events
Fixes #5.  Some songs reduce the sustain/release value to the minimum to
immediately silence a note, without the usual lingering due to the release
rate.  This is now converted as a volume event that sets the volume to zero,
which has the same effect but in a more portable way.